requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
Werkzeug==3.1.0
requests==2.31.0
pdfminer.six==20251107
# openai==1.29.0
google-genai==1.49.0
firebase-admin==6.6.0
reportlab==4.0.4